multicake/requirements.txt
asgiref==3.5.0
Django==3.2
django-modeltranslation==0.17.5
Pillow==9.0.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
